Erro ADM

+++ app/controllers/backoffice/admins_controller.rb
@@ -2,6 +2,9 @@ class Backoffice::AdminsController < BackofficeController
   
   before_action :set_admin, only: [:edit, :update, :destroy]
   
+  after_action :verify_authorized, only: [:new, :destroy]
+  after_action :verify_policy_scoped, only: :index
+
   def index
     @admins = policy_scope(Admin)
   end
@@ -15,6 +18,7 @@ class Backoffice::AdminsController < BackofficeController
     @admin = Admin.new(params_admin)
     if @admin.save	
       redirect_to backoffice_admins_path, notice: I18n.t('messages.created_with', @admin.name)
+    else
       render :new
     end
   end
@@ -26,11 +30,12 @@ class Backoffice::AdminsController < BackofficeController
     if @admin.update(params_admin)
       redirect_to backoffice_admins_path, notice: I18n.t('messages.updated_with', item: @admin.name)
     else
-      render :new
+      render :edit
     end      
   end
   
   def destroy    
+    authorize @admin
     admin_email = @admin.email
 
     if @admin.destroy
@@ -51,9 +56,9 @@ class Backoffice::AdminsController < BackofficeController
       pass_confirmation = params[:admin][:password_confirmation]
   
       if pass.blank? && pass_confirmation.blank?
-        params[:admin].except!(:password, :password_confirmation)
+        params[:admin].except(:password, :password_confirmation)
       end
       
-      params.require(:admin).permit(:name,  :email, :password, :password_confirmation)
+      params.require(:admin).permit(policy(@admin).permitted_attributes)
     end
 end
